requirements.txt
# Framework de desarrollo.
Django==3.0

# Tools to create interfaces under the REST standard.
djangorestframework==3.11.0
django-filter==2.3.0

# Library with tools for django.
django-extensions==2.2.9

#Library for loading environment variables of the file .env.
django-dotenv==1.4.2

# Debug library
ipdb==0.13.2

# library for Linting.
flake8==3.7.9

# Serialize data
marshmallow==3.7.0

# Pandas
pandas==1.0.5

# JWT
pyjwt==1.7.1

#Linting of code
pylint==2.5.3

#Sqlalchemy
sqlalchemy==1.3.18
